Make require_precursor_mz deprecated
