+++ tests/filtering/test_require_precursor_below_mz.py
@@ -28,26 +28,6 @@ def test_if_spectrum_is_cloned():
     assert not spectrum_in.get("testfield"), "Expected input spectrum to remain unchanged."
 
 
-def test_require_precursor_below_without_precursor_mz():
-    """Test if correct assert error is raised for missing precursor-mz."""
-    spectrum_in = SpectrumBuilder().build()
-
-    with pytest.raises(AssertionError) as msg:
-        _ = require_precursor_below_mz(spectrum_in)
-
-    assert str(msg.value) == "Precursor mz absent.", "Expected different error message."
-
-
-def test_require_precursor_below_with_wrong_precursor_mz():
-    """Test if correct assert error is raised for precursor-mz as string."""
-    spectrum_in = SpectrumBuilder().with_metadata({"precursor_mz": "445.0"}).build()
-
-    with pytest.raises(AssertionError) as msg:
-        _ = require_precursor_below_mz(spectrum_in)
-
-    assert "Expected 'precursor_mz' to be a scalar number." in str(msg.value)
-
-
 def test_with_input_none():
     """Test if input spectrum is None."""
     spectrum_in = None
@@ -75,11 +55,6 @@ def test_require_precursor_below_mz_max_50():
     spectrum_in = SpectrumBuilder().with_mz(mz).with_intensities(intensities).build()
     spectrum_in.set("precursor_mz", 60.)
 
-    with LogCapture() as log:
-        spectrum = require_precursor_below_mz(spectrum_in, max_mz=50)
+    spectrum = require_precursor_below_mz(spectrum_in, max_mz=50)
 
     assert spectrum is None, "Expected spectrum to be None."
-    log.check(
-        ('matchms', 'INFO', 'Spectrum with precursor_mz 60.0 (>50) was set to None.')
-    )
-    reset_matchms_logger()
