+++ tests/filtering/test_require_precursor_mz.py
@@ -34,16 +34,6 @@ def test_if_spectrum_is_cloned():
         "Expected input spectrum to remain unchanged."
 
 
-def test_require_precursor_mz_with_wrong_precursor_mz():
-    """Test if correct assert error is raised for precursor-mz as string."""
-    spectrum_in = SpectrumBuilder().with_metadata({"precursor_mz": "445.0"}).build()
-
-    with pytest.raises(AssertionError) as msg:
-        _ = require_precursor_mz(spectrum_in)
-
-    assert "Expected 'precursor_mz' to be a scalar number." in str(msg.value)
-
-
 def test_require_precursor_mz_with_input_none():
     """Test if input spectrum is None."""
     spectrum_in = None
@@ -51,23 +41,14 @@ def test_require_precursor_mz_with_input_none():
     assert spectrum is None
 
 
-@pytest.mark.parametrize("precursor_mz", [0, 10.0, -200])
+@pytest.mark.parametrize("precursor_mz", [0, 9.0, -200])
 def test_require_precursor_mz_fail_when_mz_too_small(precursor_mz):
     """Test if spectrum is None when precursor_mz <= minimum_accepted_mz"""
-    set_matchms_logger_level("INFO")
-    mz = np.array([10, 20, 30, 40], dtype="float")
-    intensities = np.array([0, 1, 10, 100], dtype="float")
-    spectrum_in = SpectrumBuilder().with_mz(mz).with_intensities(intensities).build()
+    spectrum_in = SpectrumBuilder().build()
     spectrum_in.set("precursor_mz", precursor_mz)
 
-    with LogCapture() as log:
-        spectrum = require_precursor_mz(spectrum_in)
-
+    spectrum = require_precursor_mz(spectrum_in, minimum_accepted_mz=10)
     assert spectrum is None, "Expected spectrum to be None."
-    log.check(
-        ('matchms', 'INFO', 'Spectrum without precursor_mz was set to None.')
-    )
-    reset_matchms_logger()
 
 
 def test_require_precursor_mz_fail_because_below_zero():
